Update nuxt.config, assets, pages
