+++ hashtag-dapp-nuxt/assets/css/partials/modalcard.scss
