+++ hashtag-dapp-nuxt/assets/css/partials/onboard.scss
