+++ hashtag-dapp-nuxt/assets/css/partials/widgets.scss
