+++ hashtag-dapp-nuxt/assets/css/variables.scss
