+++ hashtag-dapp-nuxt/assets/fonts/icons.otf
