+++ hashtag-dapp-nuxt/assets/fonts/outline-icons.ttf
