+++ hashtag-dapp-nuxt/assets/fonts/outline-icons.woff
