+++ hashtag-dapp-nuxt/components/MintAndTag.vue
@@ -152,9 +152,9 @@ import {
   SNAPSHOT,
   FIRST_THOUSAND_HASHTAGS,
   NFTS_ASSETS_NAME_CONTAINS,
-} from "@/queries";
+} from "~/queries";
 import { mapGetters } from "vuex";
-import HashtagValidationService from "@/services/HashtagValidationService";
+import HashtagValidationService from "~/services/HashtagValidationService";
 import debounce from "lodash/debounce";
 
 export default {
