+++ hashtag-dapp-nuxt/package.json
@@ -45,6 +45,7 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
     "eslint-plugin-vue": "^7.9.0",
+    "fibers": "^5.0.0",
     "graphql-tag": "^2.12.4",
     "html-webpack-plugin": "^5.3.1",
     "husky": "^6.0.0",
@@ -54,7 +55,8 @@
     "node-sass": "^6.0.0",
     "prettier": "^2.3.0",
     "raw-loader": "^4.0.2",
-    "sass-loader": "^11.1.0",
+    "sass": "^1.32.13",
+    "sass-loader": "10",
     "stylelint": "^13.13.1",
     "stylelint-config-recommended-scss": "^4.2.0",
     "stylelint-config-standard": "^22.0.0",
