+++ hashtag-dapp-nuxt/pages/creators/index.vue
@@ -1,11 +1,11 @@
 <template>
-  <div class="body">
+  <div class="body auction">
     <Header />
     <section class="main">
       <div class="container">
-        <h1 class="title is-1">Hashtags</h1>
+        <h1 class="title is-1">Creators</h1>
         <h2 class="subtitle">
-          Hashtag Protocol Tokens
+          Hashtag Creators
           <span class="is-pulled-right is-size-6 has-text-weight-bold">
             <router-link :to="{ name: 'dashboard' }">Dashboard</router-link
             >&nbsp;
@@ -27,37 +27,25 @@
                         <!---->
                         <th class="">
                           <div class="th-wrap">
-                            Hashtag
-                            <!---->
-                          </div>
-                        </th>
-                        <th class="">
-                          <div class="th-wrap">
-                            Created
-                            <!---->
-                          </div>
-                        </th>
-                        <th class="">
-                          <div class="th-wrap">
                             Creator
                             <!---->
                           </div>
                         </th>
                         <th class="">
-                          <div class="th-wrap">
-                            Owner
+                          <div class="th-wrap is-centered">
+                            Hashtags
                             <!---->
                           </div>
                         </th>
                         <th class="">
-                          <div class="th-wrap">
-                            Publisher
-                            <!---->
-                          </div>
-                        </th>
-                        <th class="">
-                          <div class="th-wrap">
+                          <div class="th-wrap is-centered">
                             Tag count
+                            <!---->
+                          </div>
+                        </th>
+                        <th class="">
+                          <div class="th-wrap is-centered">
+                            Revenue
                             <!---->
                           </div>
                         </th>
@@ -66,42 +54,41 @@
                       <!---->
                       <!---->
                     </thead>
-                    <tbody v-if="pagedHashtags">
+                    <tbody v-if="pagedCreators">
                       <tr
                         draggable="false"
                         class=""
-                        v-for="hashtag in pagedHashtags"
-                        v-bind:key="hashtag.id"
+                        v-for="creator in pagedCreators"
+                        v-bind:key="creator.id"
                       >
                         <!---->
                         <!---->
-                        <td data-label="Hashtag" class="">
-                          <hashtag :value="hashtag.displayHashtag"></hashtag>
+                        <td data-label="Owner" class="">
+                          <eth-account
+                            :value="creator.id"
+                            route="creator-detail"
+                          ></eth-account>
                         </td>
-                        <td data-label="Minted" class="">
-                          <timestamp-from
-                            :value="hashtag.timestamp"
-                          ></timestamp-from>
+                        <td data-label="Hashtags" class="has-text-centered">
+                          {{ creator.mintCount }}
                         </td>
-                        <td data-label="Creator" class="">
-                          <eth-account :value="hashtag.creator"></eth-account>
+                        <td data-label="Tag count" class="has-text-centered">
+                          {{ creator.tagCount }}
                         </td>
-                        <td data-label="Owner" class="">Pending Auction</td>
-                        <td data-label="Publisher" class="">
-                          <eth-account :value="hashtag.publisher"></eth-account>
-                        </td>
-                        <td data-label="Count" class="">
-                          {{ hashtag.tagCount }}
+                        <td data-label="Revenue" class="has-text-centered">
+                          <eth-amount :value="creator.tagFees"></eth-amount>
                         </td>
                         <!---->
                       </tr>
+                      <!---->
+                      <!---->
                     </tbody>
                   </table>
                 </div>
                 <!---->
               </div>
               <Pagination
-                :entity-count="hashtagCount"
+                :entity-count="creatorsCount"
                 :page-size="pageSize"
                 @tabSelected="tabSelected"
               />
@@ -115,25 +102,34 @@
 </template>
 
 <script>
-import EthAccount from "../components/EthAccount";
-import Footer from "../components/Footer";
-import Hashtag from "../components/Hashtag";
-import Header from "../components/Header";
-import Pagination from "../components/Pagination";
-import TimestampFrom from "../components/TimestampFrom";
-import { PAGED_HASHTAGS, ALL_HASHTAG_TOKEN_IDS } from "../queries";
+import EthAccount from "~/components/EthAccount";
+import EthAmount from "~/components/EthAmount";
+import Footer from "~/components/Footer";
+import Header from "~/components/Header";
+import Pagination from "~/components/Pagination";
+import { PAGED_CREATORS, ALL_CREATORS } from "~/queries";
 
 const PAGE_SIZE = 10;
 
 export default {
   name: "Nfts",
   components: {
+    EthAmount,
     EthAccount,
-    TimestampFrom,
-    Hashtag,
     Footer,
     Header,
     Pagination,
+  },
+  head() {
+    return {
+      title: '',
+      meta: [
+        {
+        hid: 'description',
+        name: 'description',
+        content: '' }
+      ],
+    }
   },
   data() {
     return {
@@ -141,12 +137,12 @@ export default {
       pageSize: PAGE_SIZE,
       first: PAGE_SIZE,
       skip: 0,
-      hashtagCount: 0,
+      creatorsCount: 0,
     };
   },
   apollo: {
-    pagedHashtags: {
-      query: PAGED_HASHTAGS,
+    pagedCreators: {
+      query: PAGED_CREATORS,
       variables() {
         return {
           first: this.first,
@@ -154,11 +150,11 @@ export default {
         };
       },
     },
-    hashtagCount: {
-      query: ALL_HASHTAG_TOKEN_IDS,
+    creatorsCount: {
+      query: ALL_CREATORS,
       manual: true,
       result({ data }) {
-        this.hashtagCount = data.hashtags.length;
+        this.creatorsCount = data.creators.length;
       },
     },
   },
