+++ hashtag-dapp-nuxt/pages/owners/index.vue
@@ -17,7 +17,7 @@
             <article class="is-white coming-soon">
               <h2 class="title is-5">Top owners</h2>
               <div class="coming-soon-img">
-                <img src="../assets/coming-soon-banner.png" />
+                <img src="../../assets/coming-soon-banner.png" />
               </div>
               <pseudo-owners />
             </article>
@@ -67,11 +67,11 @@
 </template>
 
 <script>
-import Footer from "../components/Footer";
-import Header from "../components/Header";
-import MarkdownDoc from "../components/MarkdownDoc";
-import { PAGED_OWNERS, ALL_OWNER_ADDRESSES } from "../queries";
-import PseudoOwners from "../components/PseudoOwners";
+import Footer from "../../components/Footer";
+import Header from "../../components/Header";
+import MarkdownDoc from "../../components/MarkdownDoc";
+import { PAGED_OWNERS, ALL_OWNER_ADDRESSES } from "../../queries";
+import PseudoOwners from "../../components/PseudoOwners";
 
 export default {
   name: "Nfts",
