+++ hashtag-dapp-nuxt/plugins/buefy.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import Buefy from "buefy";
+
+Vue.use(Buefy);
