+++ hashtag-dapp-nuxt/plugins/vue-axios.js
@@ -1,0 +1,5 @@
+import Vue from "vue";
+import axios from "axios";
+import VueAxios from "vue-axios";
+
+Vue.use(VueAxios, axios);
