+++ hashtag-dapp-nuxt/plugins/vue-moment.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import VueMoment from "vue-moment";
+
+Vue.use(VueMoment);
