+++ hashtag-dapp-nuxt/plugins/vue-screen.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import VueScreen from "vue-screen";
+
+Vue.use(VueScreen, "bulma");
